// src/StreamMaster.UI/lib/api.d.ts
export interface components {
	schemas: {
// ...
		M3UFileDto: components["schemas"]["BaseFileDto"] & {
			m3U8OutPutProfile: string;
			m3UName: components["schemas"]["M3UField"];
			m3UKey: components["schemas"]["M3UKey"];
			syncChannels: boolean;
			vodTags: string[];
			defaultStreamGroupName: string | null;
			/** Format: int32 */
			maxStreamCount: number;
			/** Format: int32 */
			streamCount: number;
			/** Format: int32 */
			startingChannelNumber: number;
			autoSetChannelNumbers: boolean;
		};
// ...
		PagedResponseOfM3UFileDto: components["schemas"]["DataResponseOfListOfM3UFileDto"] & {
			/** Format: int32 */
			pageNumber: number;
			/** Format: int32 */
			pageSize: number;
			/** Format: int32 */
			totalPageCount: number;
			/** Format: int32 */
			first: number;
		};
		DataResponseOfListOfM3UFileDto: components["schemas"]["APIResponse"] & {
			data: components["schemas"]["M3UFileDto"][] | null;
			/** Format: int32 */
			totalItemCount: number;
			/** Format: int32 */
			count: number;
		};
		CreateM3UFileRequest: {
			name?: string;
			/** Format: int32 */
			maxStreamCount?: number | null;
			m3U8OutPutProfile?: string | null;
			m3UKey?: components["schemas"]["M3UKey"] | null;
			m3UName?: components["schemas"]["M3UField"] | null;
			defaultStreamGroupName?: string | null;
			urlSource?: string | null;
			syncChannels?: boolean | null;
			/** Format: int32 */
			hoursToUpdate?: number | null;
			/** Format: int32 */
			startingChannelNumber?: number | null;
			autoSetChannelNumbers?: boolean | null;
			vodTags?: string[] | null;
		};
// ...
		EPGFileDto: components["schemas"]["BaseFileDto"] & {
			/** Format: int32 */
			timeShift: number;
			/** Format: int32 */
			epgNumber: number;
			color: string;
			/** Format: int32 */
			channelCount: number;
			/** Format: date-time */
			epgStartDate: string;
			/** Format: date-time */
			epgStopDate: string;
			/** Format: int32 */
			programmeCount: number;
		};
		PagedResponseOfEPGFileDto: components["schemas"]["DataResponseOfListOfEPGFileDto"] & {
			/** Format: int32 */
			pageNumber: number;
			/** Format: int32 */
			pageSize: number;
			/** Format: int32 */
			totalPageCount: number;
			/** Format: int32 */
			first: number;
		};
		DataResponseOfListOfEPGFileDto: components["schemas"]["APIResponse"] & {
			data: components["schemas"]["EPGFileDto"][] | null;
			/** Format: int32 */
			totalItemCount: number;
			/** Format: int32 */
			count: number;
		};
		CreateEPGFileRequest: {
			name?: string;
			fileName?: string;
			/** Format: int32 */
			epgNumber?: number;
			/** Format: int32 */
			timeShift?: number | null;
			/** Format: int32 */
			hoursToUpdate?: number | null;
			urlSource?: string | null;
			color?: string | null;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}